refactor: prop 타입을 prop을 사용하는 컴포넌트에서 정의하도록 수정
